devicestorage/test_enumerate_mobile.html

<!--
  Any copyright is dedicated to the Public Domain.
  http://creativecommons.org/publicdomain/zero/1.0/
-->
<!DOCTYPE HTML>
<html> <!--
https://bugzilla.mozilla.org/show_bug.cgi?id=717103
-->
<head>
  <title>Test for the device storage API </title>
  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <script type="text/javascript" src="devicestorage_common.js"></script>

<link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=717103">Mozilla Bug 717103</a>
<p id="display"></p>
<div id="content" style="display: none">
  
</div>
<pre id="test">
<script class="testbody" type="text/javascript">

SimpleTest.waitForExplicitFinish();

function enumerateSuccess(e) {

  if (e.target.result == null) {
    ok(files.length == 0, "when the enumeration is done, we shouldn't have any files in this array")
    dump("We still have length = " + files.length + "\n");
    SimpleTest.finish();
    return;
  }
  
  var filename = e.target.result.name;

  var index = files.indexOf(filename);
  files.remove(index);

  ok(index > -1, "filename should be in the enumeration : " + e.target.result.name);

  // clean up
  var cleanup = storage[0].delete(prefix + "/" + filename);
  cleanup.onsuccess = function(e) {}  // todo - can i remove this?

  e.target.continue();
}

function handleError(e) {
  ok(false, "handleError was called : " + e.target.error.name);
  SimpleTest.finish();
}

function addSuccess(e) {
  addedSoFar = addedSoFar + 1;
  if (addedSoFar == files.length) {

    var cursor = storage[0].enumerate(prefix);
    cursor.onsuccess = enumerateSuccess;
    cursor.onerror = handleError;
  }
}

function addError(e) {
  ok(false, "addError was called : " + e.target.error.name);
  SimpleTest.finish();

}

mediaTypes = ['music','movies','pictures'];

ok(navigator.getDeviceStorage, "Should have getDeviceStorage");

for (mediaType in mediaTypes) {
  var storage = navigator.getDeviceStorage(mediaTypes[mediaType]);
  
  var prefix = "devicestorage/" + randomFilename(12)
  var files = [ "a", "b", "c", "d/a", "d/b", "d/c", "d/d", "The/quick/brown/fox/jumps/over/the/lazy/dog"]
  var addedSoFar = 0;
  ok(storage, "Should be able to get the " + mediaTypes[mediaType] + " directory");
  try {
    for (var i=0; i<files.length; i++) {
      request = storage.addNamed(createRandomBlob(), prefix + '/' + files[i]);
      ok(request, "Should have a non-null request");
      request.onsuccess = addSuccess;
      request.onerror = addError;
    }
  }
  catch (e) {
    ok(false, "unknown exception called " + e);
    SimpleTest.finish();
    break;
  }
}
</script>
</pre>
</body>
</html>
